# Index modules in paths