--- a/testsuite/tests/shape-index/index_constrs.reference
+++ b/testsuite/tests/shape-index/index_constrs.reference
@@ -1,4 +1,6 @@
 Indexed shapes:
+Resolved: Index_constrs.2 :
+  M (File "index_constrs.ml", line 32, characters 32-33)
 Resolved: Index_constrs.1 :
   M.F (File "index_constrs.ml", line 32, characters 32-35)
 Resolved: Index_constrs.0 :