--- a/testsuite/tests/shape-index/index_modules.reference
+++ b/testsuite/tests/shape-index/index_modules.reference
@@ -1,4 +1,6 @@
 Indexed shapes:
+Resolved: Index_modules.1 :
+  A (File "index_modules.ml", line 22, characters 17-18)
 Resolved: Index_modules.0 :
   A.x (File "index_modules.ml", line 22, characters 17-20)
 Resolved: Index_modules.0 :