… testsuite/tests/shape-index/index.reference