Now that we have comprehensive locations stored in longidents, we can add all occurences of modules appearing in paths to the cmt's index.

This will enable external tools to provide full support for modules in project-wide features such as listing occurrences and renaming. Right now, these are restricted to direct module usages.

This PR simply adds a loop during the indexing that iterats on longidents to reduce all associated sub-path's shapes. This process is not as expensive as it appears due to the global memoization of shape reduction in the compiler. In fact, I found no meaningful difference when timing the build of the compiler:

Trunk:
```
$ hyperfine 'make world.opt –jobs 16 && (cd testsuite && make all-shape-index)' -p 'make clean' Benchmark 1: make world.opt –jobs 16 && (cd testsuite && make all-shape-index) 
Time (mean ± σ): 52.980 s ± 0.044 s [User: 242.345 s, System: 14.633 s] 
Range (min … max): 52.920 s … 53.056 s 10 runs
```

This branch:
```
hyperfine 'make world.opt –jobs 16 && (cd testsuite && make all-shape-index)' -p 'make clean' Benchmark 1: make world.opt –jobs 16 && (cd testsuite && make all-shape-index) 
Time (mean ± σ): 52.986 s ± 0.100 s [User: 242.326 s, System: 14.587 s] 
Range (min … max): 52.783 s … 53.134 s 10 runs
```

Similarly, due to sharing, the `cmt` growth is contained because we are in fact not adding new shape or location information.

Trunk:
```
$ find . -name '*.cmt*' -print0 | xargs -0  du -c       
93016   total
```

This branch:
```
$ find . -name '*.cmt*' -print0 | xargs -0  du -c        
93024   total
```